Add new custom index page: Index as a Calendar
Sometimes you want to display the index screen for a set of resources as
a calendar. To do so, use the :calendar option for the index block.

    index :as => :calendar

By default resources will be groups by their `updated_at` attribute. But
if you need to, you can specify another attribute to group by using the
:group_by argument:

    index :as => :calendar, :group_by => :published_at

The calendar is useful when showing blog posts on the day they where
written or if showing events on the day they occured.

Of cause pagination doesn't makes sense when displaying the index as a
calendar, so you should always use `config.pagiante = false` for the
resource in question.